Use python3-config by default
On my laptop, I have Python 2 installed as `python`, and therefore
its config tool is available as `python-config`. If rubicon-java
uses it to build, it yields an error because rubicon-java only
supports Python 3.

Using `python3-config` would do the right thing if you only have one
Python 3 version available, and it is already in $PATH appropriately,
so let's go with that.